Move canUseWorkspaceProject to Billing
